Add structured logging of generate actions (#531)
* Update telemetry plugin to provide entire logger as opposed to Winston config
* Refactor config ordering to ensure that telemetry is initialized before logging. This is necessary to ensure that instrumentations are loaded before the logging libraries are imported.

---------

+++ js/common/src/telemetry.ts
@@ -33,7 +33,7 @@ const actionCounter = new MetricCounter(_N('requests'), {
 
 const actionLatencies = new MetricHistogram(_N('action_latency'), {
   description: 'Latencies when calling Genkit actions.',
-  valueType: ValueType.INT,
+  valueType: ValueType.DOUBLE,
   unit: 'ms',
 });
